# Allowing case insensitive regexp groups after PR #144