Changes in https://github.com/gorilla/mux/pull/144 broke regexp matching for case insensitive groups. 

This adds a test scenario and a small hunk of code reverted to what it was prior to the PR.