@jprobinson 2016-03-07
Benchmark changes: 
```
benchmark                             old ns/op     new ns/op     delta
BenchmarkMux-4                        2712          2569          -5.27%
BenchmarkMuxAlternativeInRegexp-4     3835          3860          +0.65%
BenchmarkManyPathVariables-4          4410          4476          +1.50%
```