@elithrar 2016-03-07
Thanks @jprobinson—and thanks for adding the benchmarks (which I was just about to ask for). Happy to lose a few % if it means we maintain API compat, especially since request muxes are almost never bottlenecks in real applications.